factory/requirements.txt
cairocffi==1.0.2
CairoSVG==2.4.0
cffi==1.12.3
cssselect2==0.2.1
defusedxml==0.6.0
html5lib==1.0.1
lxml==4.3.4
Markdown==3.1.1
pdfrw==0.4
Pillow==6.0.0
pycparser==2.19
Pygments==2.4.2
Pyphen==0.9.5
six==1.12.0
tinycss2==1.0.2
WeasyPrint==0.42.3
webencodings==0.5.1
